fix: suppress test output
