+++ test/progressTracker.test.js
@@ -3,101 +3,10 @@
 const helper = require('./helper')
 const test = require('tap').test
 const { defaultMaxListeners } = require('events')
+const { Writable } = require('stream')
 const sinon = require('sinon')
-const progressTracker = require('../lib/progressTracker')
 const autocannon = require('../autocannon')
 const { hasWorkerSupport } = require('../lib/util')
-
-test('progress tracker should throw if no instance is provided', t => {
-  t.plan(1)
-  try {
-    progressTracker(null, {})
-  } catch (error) {
-    t.same(error.message, 'instance required for tracking')
-  }
-})
-
-test('should work', t => {
-  const server = helper.startServer({ statusCode: 404 })
-  const instance = autocannon({
-    url: `http://localhost:${server.address().port}`,
-    pipelining: 2
-  }, console.log)
-
-  setTimeout(() => {
-    instance.stop()
-    t.end()
-  }, 2000)
-
-  autocannon.track(instance, {
-    renderProgressBar: true,
-    renderLatencyTable: true
-  })
-})
-
-test('should work with amount', t => {
-  const server = helper.startServer()
-  const instance = autocannon({
-    url: `http://localhost:${server.address().port}`,
-    pipelining: 1,
-    amount: 10
-  }, process.stdout)
-
-  setTimeout(() => {
-    instance.stop()
-    t.end()
-  }, 2000)
-  autocannon.track(instance, {
-    renderProgressBar: true
-  })
-  t.pass()
-})
-
-test('should log mismatches', t => {
-  const server = helper.startServer()
-  const instance = autocannon({
-    url: `http://localhost:${server.address().port}`,
-    pipelining: 1,
-    amount: 10,
-    expectBody: 'modified'
-  }, console.log)
-
-  setTimeout(() => {
-    instance.stop()
-    t.end()
-  }, 2000)
-  autocannon.track(instance, {
-    renderProgressBar: true
-  })
-  t.pass()
-})
-
-test('should log resets', t => {
-  const server = helper.startServer()
-  const instance = autocannon({
-    url: `http://localhost:${server.address().port}`,
-    connections: 1,
-    amount: 10,
-    requests: [
-      { method: 'GET' },
-      {
-        method: 'GET',
-        // falsey result will reset
-        setupRequest: () => {}
-      },
-      { method: 'GET' }
-    ]
-  }, console.log)
-
-  setTimeout(() => {
-    instance.stop()
-    t.end()
-  }, 2000)
-  autocannon.track(instance, {
-    renderProgressBar: true
-  })
-  t.pass()
-})
 
 test(`should not emit warnings when using >= ${defaultMaxListeners} workers`, { skip: !hasWorkerSupport }, t => {
   const server = helper.startServer()
@@ -117,5 +26,9 @@ test(`should not emit warnings when using >= ${defaultMaxListeners} workers`, { 
 
   const emitWarningSpy = sinon.spy(process, 'emitWarning')
 
-  autocannon.track(instance)
+  autocannon.track(instance, {
+    outputStream: new Writable({
+      write () {}
+    })
+  })
 })
